requirements.txt
Flask
Flask-PyMongo
python-dotenv
pandas
openpyxl
